==> 1. Data Exploration in Spark.ipynb <==
Step 1:
Load data into Spark DataFrame
We will need to import the SQLContext class

from pyspark.sql import SQLContext

We will create an SQLContext

sqlContext = SQLContext(sc)

Read the weather data into a DataFrame

The FIRST arguement specifies the URL to the daily_weather .csv file
The SECOND arguement specifies the spark-csv format
The THIRD arguement says the first line in daily_weather is the header
The FOURTH arguement says to infer the data types

df = sqlContext.read.load('file:///home/cloudera/Downloads/big-data-4/daily_weather.csv',
                         format='com.databricks.spark.csv',
                         header='true',inferSchema='true')

Step 2:
Look at data columns and types by running df.columns    

df.columns

We can view the data type for each column by calling printSchema()

df.printSchema()

Step 3:
Print Summary Statistics
Print summary statistics for all columns using describe() method 

df.describe().toPandas().transpose()

We can also see the summary for just one column:

df.describe('air_pressure_9am').show()

Let's count the number of columns and rows in the DataFrame:

len(df.columns)

Count the rows

df.count()

Step 4:
Drop rows with missing value
Let's drop the rows with missing values in the air_pressure_9am columns

df2 = df.na.drop(subset=['air_pressure_9am'])

Let's see the total number of row after dropping missing values

df2.count()

Step 5:
Compute the correlation between two columns
Using corr() method to compute the correlation between two columns.
Let's compute the correlation between rain_acculumation_9am and rain_duration_9am:

df2.stat.corr('rain_accumulation_9am','rain_duration_9am')
