==> 2. Handling Missing Values in Spark.ipynb <==
Step 1:
Load classes and weather data

Run the cell to:
-load the SQLContext, 
-create instance of SQLContext and
-read the weather data into a DataFrame

from pyspark.sql import SQLContext

sqlContext = SQLContext(sc)

df = sqlContext.read.load('file:///home/cloudera/Downloads/big-data-4/daily_weather.csv', 
                          format='com.databricks.spark.csv', 
                          header='true',inferSchema='true')

Step 2:
Print the summary statistics for ALL COLUMNS using describe()

df.describe().toPandas().transpose()

Let's just look at the statistics for the air temp at 9am:

df.describe(['air_temp_9am']).show()

Notice that the total number of rows in the DataFrame is 1095 while the number of rows in air_temp_9am is 1090.

This indicates that there are missing values in air_temp_9am

The following command shows the total number of rows in the DataFrame:

df.count()

Step 3:
Remove missing values

We can drop all the rows missing a value in any calling using na.drop()

removeAllDF = df.na.drop()

Let's look at the summary statistics for air_temp_9am with the MISSING VALUE DROPPED:

removeAllDF.describe(['air_temp_9am']).show()

Now we have to compare the summary statistics before and after dropping the missing value

We can see that the mean and standard deviation is close to the original values, for mean is 64.933 vs. 65.022 and standard deviation is 11.175 vs. 11.168

Also you can notice that the count is 1064, which means that 1095-1064 = 31 rows dropped.

We can see this agrees with the total number of rows in the new DataFrame:

removeAllDF.count()

By far the process goes like:

Read the file into DataFrame -> 
Print summary statistics ->
Look at statistics for air_temp_9am ->
Remove missing values ->
Look at statistics for air_temp_9am after dropping->
Look at number of rows for DataFrame

Step 4:
Impute missing values

Instead of removing rows containing missing values,
let's replace the values with the mean value for that column.

FIRST, we'll load the avg function and make a COPY of the original DataFrame for future comparision

from pyspark.sql.functions import avg
imputeDF = df

Next, we'll iterate through each column in the DataFrame:

compute the mean value fot that COLUMN and replace any MISSING VALUES in that column with the mean.

for x in imputeDF.columns:
    meanValue = removeAllDF.agg(avg(x)).first()[0]
    print(x, meanValue)
    imputeDF = imputeDF.na.fill(meanValue, [x])

The agg() function performs an aggregate calculation on the DataFrame and avg(x) specifies to compute the mean on column x.

The agg() function returns a DataFrame, first() returns the FIRST row, and [0] gets the first value.

The LAST line of the code uses na.fill() to replace the missing values with the mean value(FIRST ARGUEMENT) in column x(SECOND ARGUEMENT).

The output of executing this cell prints the mean values for each column and we can see the mean value for air_temp_9am is the same mean when we remove all the missing values in step 4, i.e., 65.022

# Either we remove the missing value or replace them with mean, we get the same result

STEP 5:
Print imputed data summary statistics

Let's call describe() to show the summary statistics for the ORIGINAL and IMPUTED air_temp_9am

df.describe(['air_temp_9am']).show()
imputeDF.describe(['air_temp_9am']).show()
